+++ server/config/config.js
@@ -1,0 +1,5 @@
+const TOKEN_SECRET = "some secret key";
+
+module.exports = {
+  TOKEN_SECRET,
+};
